Issue solved: can not import a module from another module Standardize the member vars `api_mod_paths` and `api_lib_paths` as maps

+++ ports/W801/api/leds/write.c.v
@@ -10,7 +10,7 @@ import pin
 
 // write a integer value to the builtin LEDs
 pub fn write(value u8) {
-	mask := 0b00000001
+	mask := u8(0b00000001)
 	for n in 0..6 {
 		pin.write( led_array[n], ( value & (mask << n) ) >> n)
 	}
